requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
cryptography==44.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.6
django-admin-interface==0.29.4
django-admin-tools==0.9.3
django-allauth==65.5.0
django-colorfield @ git+https://github.com/fabiocaccamo/django-colorfield.git@ff121f6ff3ef577d79c7b25ac2f24120539c6d77
django-cors-headers==4.7.0
django-environ==0.12.0
django-grappelli==4.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Flask==3.1.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
MarkupSafe==3.0.2
marshmallow==3.26.0
npm==0.1.1
optional-django==0.1.0
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.9.0
pypdf==5.4.0
pypiwin32==223
python-bidi==0.6.6
python-decouple==3.8
python-slugify==8.0.4
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.17
psycopg2>=2.9,<3.0
dj-database-url>=0.5,<1.0
